Module arecibo::provider

This module implements Nova’s traits using the following several different combinations

Modules

hyperkzg

This module implements Nova’s evaluation engine using HyperKZG, a KZG-based polynomial commitment for multilinear polynomials HyperKZG is based on the transformation from univariate PCS to multilinear PCS in the Gemini paper (section 2.4.2 in <https://eprint.iacr.org/2022/420.pdf>). However, there are some key differences: (1) HyperKZG works with multilinear polynomials represented in evaluation form (rather than in coefficient form in Gemini’s transformation). This means that Spartan’s polynomial IOP can use commit to its polynomials as-is without incurring any interpolations or FFTs. (2) HyperKZG is specialized to use KZG as the univariate commitment scheme, so it includes several optimizations (both during the transformation of multilinear-to-univariate claims and within the KZG commitment scheme implementation itself).

ipa_pc

This module implements EvaluationEngine using an IPA-based polynomial commitment scheme

non_hiding_zeromorph

Non-hiding Zeromorph scheme for Multilinear Polynomials.

shplonk

Shplonk PCS

Structs

Bn256Engine

An implementation of the Nova Engine trait with BN254 curve and Pedersen commitment scheme

Bn256EngineKZG

An implementation of Nova traits with multilinear KZG over the BN256 curve

Bn256EngineZM

An implementation of the Nova Engine trait with BN254 curve and Zeromorph commitment scheme

GrumpkinEngine

An implementation of the Nova Engine trait with Grumpkin curve and Pedersen commitment scheme

PallasEngine

An implementation of the Nova Engine trait with Pallas curve and Pedersen commitment scheme

Secp256k1Engine

An implementation of the Nova Engine trait with Secp256k1 curve and Pedersen commitment scheme

Secq256k1Engine

An implementation of the Nova Engine trait with Secp256k1 curve and Pedersen commitment scheme

VestaEngine

An implementation of the Nova Engine trait with Vesta curve and Pedersen commitment scheme