Module arecibo::provider::non_hiding_zeromorph

Non-hiding Zeromorph scheme for Multilinear Polynomials.

Structs

UVKZGEvaluation

Polynomial Evaluation

UVKZGPCS

KZG Polynomial Commitment Scheme on univariate polynomial. Note: this is non-hiding, which is why we will implement traits on this token struct, as we expect to have several impls for the trait pegged on the same instance of a pairing::Engine.

UVKZGProof

Proofs

ZMCommitment

Commitments

ZMEvaluation

Polynomial Evaluation

ZMPCS

Zeromorph Polynomial Commitment Scheme on multilinear polynomials. Note: this is non-hiding, which is why we will implement the EvaluationEngineTrait on this token struct, as we will have several impls for the trait pegged on the same instance of a pairing::Engine.

ZMProof

Proofs

ZMProverKey

ZMProverKey is used to generate a proof

ZMVerifierKey

ZMVerifierKey is used to check evaluation proofs for a given commitment.

Type Aliases

UVKZGPoly

Polynomial and its associated types