Module arecibo::spartan

This module implements RelaxedR1CSSNARKTrait using Spartan that is generic over the polynomial commitment and evaluation argument (i.e., a PCS) We provide two implementations, one in snark.rs (which does not use any preprocessing) and another in ppsnark.rs (which uses preprocessing to keep the verifier’s state small if the PCS provides a succinct verifier) We also provide direct.rs that allows proving a step circuit directly with either of the two SNARKs.

In polynomial.rs we also provide foundational types and functions for manipulating multilinear polynomials.

Modules

batched

This module implements BatchedRelaxedR1CSSNARKTrait using Spartan that is generic over the polynomial commitment and evaluation argument (i.e., a PCS) This version of Spartan does not use preprocessing so the verifier keeps the entire description of R1CS matrices. This is essentially optimal for the verifier when using an IPA-based polynomial commitment scheme. This batched implementation batches the outer and inner sumchecks of the Spartan SNARK.

batched_ppsnark

batched pp snark

polys

This module contains the definitions of polynomial types used in the Spartan SNARK.

ppsnark

This module implements RelaxedR1CSSNARK traits using a spark-based approach to prove evaluations of sparse multilinear polynomials involved in Spartan’s sum-check protocol, thereby providing a preprocessing SNARK The verifier in this preprocessing SNARK maintains a commitment to R1CS matrices. This is beneficial when using a polynomial commitment scheme in which the verifier’s costs is succinct. This code includes experimental optimizations to reduce runtimes and proof sizes.

snark

This module implements RelaxedR1CSSNARKTrait using Spartan that is generic over the polynomial commitment and evaluation argument (i.e., a PCS) This version of Spartan does not use preprocessing so the verifier keeps the entire description of R1CS matrices. This is essentially optimal for the verifier when using an IPA-based polynomial commitment scheme.