Module arecibo::spartan::batched

This module implements BatchedRelaxedR1CSSNARKTrait using Spartan that is generic over the polynomial commitment and evaluation argument (i.e., a PCS) This version of Spartan does not use preprocessing so the verifier keeps the entire description of R1CS matrices. This is essentially optimal for the verifier when using an IPA-based polynomial commitment scheme. This batched implementation batches the outer and inner sumchecks of the Spartan SNARK.

Structs

BatchedRelaxedR1CSSNARK

A succinct proof of knowledge of a witness to a batch of relaxed R1CS instances The proof is produced using Spartan’s combination of the sum-check and the commitment to a vector viewed as a polynomial commitment

ProverKey

A type that represents the prover’s key

VerifierKey

A type that represents the verifier’s key