Module arecibo::spartan::ppsnark

This module implements RelaxedR1CSSNARK traits using a spark-based approach to prove evaluations of sparse multilinear polynomials involved in Spartan’s sum-check protocol, thereby providing a preprocessing SNARK The verifier in this preprocessing SNARK maintains a commitment to R1CS matrices. This is beneficial when using a polynomial commitment scheme in which the verifier’s costs is succinct. This code includes experimental optimizations to reduce runtimes and proof sizes.

Structs

ProverKey

A type that represents the prover’s key

R1CSShapeSparkCommitment

A type that holds a commitment to a sparse polynomial

R1CSShapeSparkRepr

A type that holds R1CSShape in a form amenable to memory checking

RelaxedR1CSSNARK

A succinct proof of knowledge of a witness to a relaxed R1CS instance The proof is produced using Spartan’s combination of the sum-check and the commitment to a vector viewed as a polynomial commitment

VerifierKey

A type that represents the verifier’s key