Trait arecibo::supernova::NonUniformCircuit

pub trait NonUniformCircuit<E1>
where
    E1: CurveCycleEquipped,
{
    type C1: StepCircuit<E1::Scalar>;
    type C2: StepCircuit<<Dual<E1> as Engine>::Scalar>;

    // Required methods
    fn num_circuits(&self) -> usize;
    fn primary_circuit(&self, circuit_index: usize) -> Self::C1;
    fn secondary_circuit(&self) -> Self::C2;

    // Provided method
    fn initial_circuit_index(&self) -> usize { ... }
}

SuperNova helper trait, for implementors that provide sets of sub-circuits to be proved via NIVC. C1 must be a type (likely an Enum) for which a potentially-distinct instance can be supplied for each index below self.num_circuits().

Required Associated Types

type C1: StepCircuit<E1::Scalar>

The type of the step-circuits on the primary

type C2: StepCircuit<<Dual<E1> as Engine>::Scalar>

The type of the step-circuits on the secondary

Required Methods

fn num_circuits(&self) -> usize

How many circuits are provided?

fn primary_circuit(&self, circuit_index: usize) -> Self::C1

Return a new instance of the primary circuit at index.

fn secondary_circuit(&self) -> Self::C2

Return a new instance of the secondary circuit.

Provided Methods

fn initial_circuit_index(&self) -> usize

Initial circuit index, defaults to zero.