Module arecibo::traits

This module defines various traits required by the users of the library to implement.

Modules

circuit

This module defines traits that a step function must implement

commitment

This module defines a collection of traits that define the behavior of a commitment engine We require the commitment engine to provide a commitment to vectors with a single group element

evaluation

This module defines a collection of traits that define the behavior of a polynomial evaluation engine A vector of size N is treated as a multilinear polynomial in \log{N} variables, and a commitment provided by the commitment engine is treated as a multilinear polynomial commitment

snark

This module defines a collection of traits that define the behavior of a zkSNARK for RelaxedR1CS

Traits

AbsorbInROTrait

A helper trait to absorb different objects in RO

CurveCycleEquipped

This is a convenience trait to pair engines which fields are in a curve cycle relationship

Engine

A collection of engines that are required by the library

Group

Represents an element of a group This is currently tailored for an elliptic curve group

PrimeFieldExt

Defines additional methods on PrimeField objects

ROCircuitTrait

A helper trait that defines the behavior of a hash function that we use as an RO in the circuit model

ROTrait

A helper trait that defines the behavior of a hash function that we use as an RO

TranscriptEngineTrait

This trait defines the behavior of a transcript engine compatible with Spartan

TranscriptReprTrait

This trait allows types to implement how they want to be added to TranscriptEngine

Type Aliases

Dual

Convenience projection to the secondary Engine of a CurveCycleEquipped

ROConstants

An alias for constants associated with E::RO

ROConstantsCircuit

An alias for constants associated with E::ROCircuit