Trait arecibo::traits::commitment::CommitmentEngineTrait

pub trait CommitmentEngineTrait<E: Engine>: Clone + Send + Sync {
    type CommitmentKey: Len + Clone + PartialEq + Debug + Send + Sync + Serialize + for<'de> Deserialize<'de> + Abomonation;
    type Commitment: CommitmentTrait<E>;

    // Required methods
    fn setup(label: &'static [u8], n: usize) -> Self::CommitmentKey;
    fn commit(ck: &Self::CommitmentKey, v: &[E::Scalar]) -> Self::Commitment;
}

A trait that ties different pieces of the commitment generation together

Required Associated Types

type CommitmentKey: Len + Clone + PartialEq + Debug + Send + Sync + Serialize + for<'de> Deserialize<'de> + Abomonation

Holds the type of the commitment key The key should quantify its length in terms of group generators.

type Commitment: CommitmentTrait<E>

Holds the type of the commitment

Required Methods

fn setup(label: &'static [u8], n: usize) -> Self::CommitmentKey

Samples a new commitment key of a specified size

fn commit(ck: &Self::CommitmentKey, v: &[E::Scalar]) -> Self::Commitment

Commits to the provided vector using the provided generators

Object Safety

This trait is not object safe.