Trait arecibo::traits::evaluation::EvaluationEngineTrait

pub trait EvaluationEngineTrait<E: Engine>: Clone + Send + Sync {
    type ProverKey: Send + Sync;
    type VerifierKey: Send + Sync + Serialize;
    type EvaluationArgument: Clone + Send + Sync + Serialize + for<'de> Deserialize<'de>;

    // Required methods
    fn setup(
        ck: Arc<<<E as Engine>::CE as CommitmentEngineTrait<E>>::CommitmentKey>
    ) -> (Self::ProverKey, Self::VerifierKey);
    fn prove(
        ck: &<<E as Engine>::CE as CommitmentEngineTrait<E>>::CommitmentKey,
        pk: &Self::ProverKey,
        transcript: &mut E::TE,
        comm: &<<E as Engine>::CE as CommitmentEngineTrait<E>>::Commitment,
        poly: &[E::Scalar],
        point: &[E::Scalar],
        eval: &E::Scalar
    ) -> Result<Self::EvaluationArgument, NovaError>;
    fn verify(
        vk: &Self::VerifierKey,
        transcript: &mut E::TE,
        comm: &<<E as Engine>::CE as CommitmentEngineTrait<E>>::Commitment,
        point: &[E::Scalar],
        eval: &E::Scalar,
        arg: &Self::EvaluationArgument
    ) -> Result<(), NovaError>;
}

A trait that ties different pieces of the commitment evaluation together

Required Associated Types

type ProverKey: Send + Sync

A type that holds the prover key

type VerifierKey: Send + Sync + Serialize

A type that holds the verifier key

type EvaluationArgument: Clone + Send + Sync + Serialize + for<'de> Deserialize<'de>

A type that holds the evaluation argument

Required Methods

fn setup( ck: Arc<<<E as Engine>::CE as CommitmentEngineTrait<E>>::CommitmentKey> ) -> (Self::ProverKey, Self::VerifierKey)

A method to perform any additional setup needed to produce proofs of evaluations

Note: This method should be cheap and should not copy most of the commitment key. Look at CommitmentEngineTrait::setup for generating SRS data.

fn prove( ck: &<<E as Engine>::CE as CommitmentEngineTrait<E>>::CommitmentKey, pk: &Self::ProverKey, transcript: &mut E::TE, comm: &<<E as Engine>::CE as CommitmentEngineTrait<E>>::Commitment, poly: &[E::Scalar], point: &[E::Scalar], eval: &E::Scalar ) -> Result<Self::EvaluationArgument, NovaError>

A method to prove the evaluation of a multilinear polynomial

fn verify( vk: &Self::VerifierKey, transcript: &mut E::TE, comm: &<<E as Engine>::CE as CommitmentEngineTrait<E>>::Commitment, point: &[E::Scalar], eval: &E::Scalar, arg: &Self::EvaluationArgument ) -> Result<(), NovaError>

A method to verify the purported evaluation of a multilinear polynomials

Object Safety

This trait is not object safe.