Function arecibo::traits::snark::default_ck_hint

pub fn default_ck_hint<E: Engine>(
) -> Box<dyn for<'a> Fn(&'a R1CSShape<E>) -> usize>

Public parameter creation takes a size hint. This size hint carries the particular requirements of the final compressing SNARK the user expected to use with these public parameters, and the below is a sensible default, which is to not require any more bases then the usual (maximum of the number of variables and constraints of the involved R1CS circuit).