Trait arecibo::traits::Engine

pub trait Engine: Clone + Copy + Debug + Send + Sync + Sized + Eq + PartialEq {
    type Base: PrimeFieldBits + TranscriptReprTrait<Self::GE> + Serialize + for<'de> Deserialize<'de>;
    type Scalar: PrimeFieldBits + PrimeFieldExt + Send + Sync + TranscriptReprTrait<Self::GE> + Serialize + for<'de> Deserialize<'de>;
    type GE: Group<Base = Self::Base, Scalar = Self::Scalar> + Serialize + for<'de> Deserialize<'de>;
    type RO: ROTrait<Self::Base, Self::Scalar>;
    type ROCircuit: ROCircuitTrait<Self::Base>;
    type TE: TranscriptEngineTrait<Self>;
    type CE: CommitmentEngineTrait<Self>;
}

A collection of engines that are required by the library

Required Associated Types

type Base: PrimeFieldBits + TranscriptReprTrait<Self::GE> + Serialize + for<'de> Deserialize<'de>

A type representing an element of the base field of the group

type Scalar: PrimeFieldBits + PrimeFieldExt + Send + Sync + TranscriptReprTrait<Self::GE> + Serialize + for<'de> Deserialize<'de>

A type representing an element of the scalar field of the group

type GE: Group<Base = Self::Base, Scalar = Self::Scalar> + Serialize + for<'de> Deserialize<'de>

A type that represents an element of the group

type RO: ROTrait<Self::Base, Self::Scalar>

A type that represents a circuit-friendly sponge that consumes elements from the base field and squeezes out elements of the scalar field

type ROCircuit: ROCircuitTrait<Self::Base>

An alternate implementation of Self::RO in the circuit model

type TE: TranscriptEngineTrait<Self>

A type that provides a generic Fiat-Shamir transcript to be used when externalizing proofs

type CE: CommitmentEngineTrait<Self>

A type that defines a commitment engine over scalars in the group

Object Safety

This trait is not object safe.