Trait arecibo::traits::Group

pub trait Group: Clone + Copy + Debug + Send + Sync + Sized + Eq + PartialEq {
    type Base: PrimeFieldBits + Serialize + for<'de> Deserialize<'de>;
    type Scalar: PrimeFieldBits + PrimeFieldExt + Send + Sync + Serialize + for<'de> Deserialize<'de>;

    // Required method
    fn group_params() -> (Self::Base, Self::Base, BigInt, BigInt);
}

Represents an element of a group This is currently tailored for an elliptic curve group

Required Associated Types

type Base: PrimeFieldBits + Serialize + for<'de> Deserialize<'de>

A type representing an element of the base field of the group

type Scalar: PrimeFieldBits + PrimeFieldExt + Send + Sync + Serialize + for<'de> Deserialize<'de>

A type representing an element of the scalar field of the group

Required Methods

fn group_params() -> (Self::Base, Self::Base, BigInt, BigInt)

Returns A, B, the order of the group, the size of the base field as big integers