Trait arecibo::traits::TranscriptReprTrait

pub trait TranscriptReprTrait<G: Group>: Send + Sync {
    // Required method
    fn to_transcript_bytes(&self) -> Vec<u8> ;
}

This trait allows types to implement how they want to be added to TranscriptEngine

Required Methods

fn to_transcript_bytes(&self) -> Vec<u8>

returns a byte representation of self to be added to the transcript