Trait arecibo::traits::snark::RelaxedR1CSSNARKTrait

pub trait RelaxedR1CSSNARKTrait<E: Engine>: Send + Sync + Serialize + for<'de> Deserialize<'de> {
    type ProverKey: Send + Sync;
    type VerifierKey: Send + Sync + Serialize;

    // Required methods
    fn setup(
        ck: Arc<<<E as Engine>::CE as CommitmentEngineTrait<E>>::CommitmentKey>,
        S: &R1CSShape<E>
    ) -> Result<(Self::ProverKey, Self::VerifierKey), NovaError>;
    fn prove(
        ck: &<<E as Engine>::CE as CommitmentEngineTrait<E>>::CommitmentKey,
        pk: &Self::ProverKey,
        S: &R1CSShape<E>,
        U: &RelaxedR1CSInstance<E>,
        W: &RelaxedR1CSWitness<E>
    ) -> Result<Self, NovaError>;
    fn verify(
        &self,
        vk: &Self::VerifierKey,
        U: &RelaxedR1CSInstance<E>
    ) -> Result<(), NovaError>;

    // Provided method
    fn ck_floor() -> Box<dyn for<'a> Fn(&'a R1CSShape<E>) -> usize> { ... }
}

A trait that defines the behavior of a zkSNARK

Required Associated Types

type ProverKey: Send + Sync

A type that represents the prover’s key

type VerifierKey: Send + Sync + Serialize

A type that represents the verifier’s key

Required Methods

fn setup( ck: Arc<<<E as Engine>::CE as CommitmentEngineTrait<E>>::CommitmentKey>, S: &R1CSShape<E> ) -> Result<(Self::ProverKey, Self::VerifierKey), NovaError>

Produces the keys for the prover and the verifier

fn prove( ck: &<<E as Engine>::CE as CommitmentEngineTrait<E>>::CommitmentKey, pk: &Self::ProverKey, S: &R1CSShape<E>, U: &RelaxedR1CSInstance<E>, W: &RelaxedR1CSWitness<E> ) -> Result<Self, NovaError>

Produces a new SNARK for a relaxed R1CS

fn verify( &self, vk: &Self::VerifierKey, U: &RelaxedR1CSInstance<E> ) -> Result<(), NovaError>

Verifies a SNARK for a relaxed R1CS

Provided Methods

fn ck_floor() -> Box<dyn for<'a> Fn(&'a R1CSShape<E>) -> usize>

This associated function (not a method) provides a hint that offers a minimum sizing cue for the commitment key used by this SNARK implementation. The commitment key passed in setup should then be at least as large as this hint.

Object Safety

This trait is not object safe.